Migrate from Qt 5.13.2 to Qt 6.8.3 LTS with Schannel TLS
- Update all CMakeLists.txt: Qt5:: -> Qt6::, Qt6 DLL names, Schannel TLS plugin

- Fix Qt 6 API changes: QNetworkReply::errorOccurred, QString::asprintf, QJsonValue::toObject().value(), QVariant(long) cast, modernize SIGNAL/SLOT

- Add missing includes exposed by Qt 6 reduced transitive headers: GL/glew.h (manager.h), ostream (animated.h, WoWModel.cpp), using std::endl (WoWModel.cpp)

- Remove _HAS_STD_BYTE=0 from cmake/windows.cmake (Qt 6 requires std::byte)

- Remove Qt 5.13.2 from git tracking (9,256 files), download via aqtinstall in CI

- Remove OpenSSL 1.1.1n (4 files), replaced by native Windows Schannel TLS backend

- Remove unused TGA image format plugin (not in Qt 6 base, not used by project)

- Remove unused x64-windows vcpkg triplet overlay (was only for OpenSSL)

- Remove libjpeg-turbo from vcpkg deps (removed in prior commit)

- Update setup_x64_deps.ps1 for Qt 6.8.3 and cleaned dependencies

- Add ThirdParty/Qt/ to .gitignore (downloaded on demand by CI and locally)

+++ Source/wowmodelviewer/app.cpp
@@ -348,7 +348,7 @@ void WowModelViewApp::SaveSettings()
 	// Application Config Settings
 	QSettings config(QString::fromWCharArray(cfgPath.c_str()), QSettings::IniFormat);
 
-	config.setValue("Locale/LanguageID", langID);
+	config.setValue("Locale/LanguageID", static_cast<int>(langID));
 	config.setValue("Locale/LanguageName", QString::fromWCharArray(langName.c_str()));
 
 	config.setValue("Settings/Path", QString::fromWCharArray(gamePath.c_str()));
